requirements.txt
webvtt-py
yt-dlp
youtube-search-python
